+++ tests/installation/test_installer.py
@@ -678,11 +678,12 @@ def test_run_installs_with_local_file(installer, locker, repo, package):
 def test_run_installs_with_local_poetry_directory_and_extras(
     installer, locker, repo, package, tmpdir
 ):
-    file_path = Path("tests/fixtures/project_with_extras/")
-    package.add_dependency("demo", {"path": str(file_path), "extras": ["extras_a"]})
+    file_path = Path("tests/fixtures/project_with_extras")
+    package.add_dependency(
+        "project-with-extras", {"path": str(file_path), "extras": ["extras_a"]}
+    )
 
     repo.add_package(get_package("pendulum", "1.4.4"))
-    repo.add_package(get_package("cachy", "0.2.0"))
 
     installer.run()
 
@@ -699,7 +700,9 @@ def test_run_installs_with_local_poetry_directory_transitive(
     file_path = Path(
         "tests/fixtures/directory/project_with_transitive_directory_dependencies/"
     )
-    package.add_dependency("demo", {"path": str(file_path)})
+    package.add_dependency(
+        "project-with-transitive-directory-dependencies", {"path": str(file_path)}
+    )
 
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cachy", "0.2.0"))
@@ -717,7 +720,7 @@ def test_run_installs_with_local_setuptools_directory(
     installer, locker, repo, package, tmpdir
 ):
     file_path = Path("tests/fixtures/project_with_setup/")
-    package.add_dependency("demo", {"path": str(file_path)})
+    package.add_dependency("my-package", {"path": str(file_path)})
 
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cachy", "0.2.0"))