+++ tests/puzzle/test_solver.py
@@ -10,6 +10,7 @@ from poetry.repositories.pool import Pool
 from poetry.repositories.repository import Repository
 from poetry.puzzle import Solver
 from poetry.puzzle.exceptions import SolverProblemError
+from poetry.utils._compat import Path
 from poetry.version.markers import parse_marker
 
 from tests.helpers import get_dependency
@@ -913,7 +914,6 @@ def test_solver_can_resolve_git_dependencies(solver, repo, package):
         ops,
         [
             {"job": "install", "package": pendulum},
-            {"job": "install", "package": cleo},
             {"job": "install", "package": get_package("demo", "0.1.2")},
         ],
     )
@@ -1246,7 +1246,6 @@ def test_solver_git_dependencies_update(solver, repo, package, installed):
         ops,
         [
             {"job": "install", "package": pendulum},
-            {"job": "install", "package": cleo},
             {
                 "job": "update",
                 "from": get_package("demo", "0.1.2"),
@@ -1255,9 +1254,10 @@ def test_solver_git_dependencies_update(solver, repo, package, installed):
         ],
     )
 
-    op = ops[2]
+    op = ops[1]
 
     assert op.job_type == "update"
+    assert op.package.source_type == "git"
     assert op.package.source_reference.startswith("9cf87a2")
     assert op.initial_package.source_reference == "123456"
 
@@ -1282,7 +1282,6 @@ def test_solver_git_dependencies_update_skipped(solver, repo, package, installed
         ops,
         [
             {"job": "install", "package": pendulum},
-            {"job": "install", "package": cleo},
             {
                 "job": "install",
                 "package": get_package("demo", "0.1.2"),
@@ -1290,3 +1289,202 @@ def test_solver_git_dependencies_update_skipped(solver, repo, package, installed
             },
         ],
     )
+
+
+def test_solver_can_resolve_directory_dependencies(solver, repo, package):
+    pendulum = get_package("pendulum", "2.0.3")
+    repo.add_package(pendulum)
+
+    path = str(
+        Path(__file__).parent.parent
+        / "fixtures"
+        / "git"
+        / "github.com"
+        / "demo"
+        / "demo"
+    )
+
+    package.add_dependency("demo", {"path": path})
+
+    ops = solver.solve()
+
+    check_solver_result(
+        ops,
+        [
+            {"job": "install", "package": pendulum},
+            {"job": "install", "package": get_package("demo", "0.1.2")},
+        ],
+    )
+
+    op = ops[1]
+
+    assert op.package.name == "demo"
+    assert op.package.version.text == "0.1.2"
+    assert op.package.source_type == "directory"
+    assert op.package.source_url == path
+
+
+def test_solver_can_resolve_directory_dependencies_with_extras(solver, repo, package):
+    pendulum = get_package("pendulum", "2.0.3")
+    cleo = get_package("cleo", "1.0.0")
+    repo.add_package(pendulum)
+    repo.add_package(cleo)
+
+    path = str(
+        Path(__file__).parent.parent
+        / "fixtures"
+        / "git"
+        / "github.com"
+        / "demo"
+        / "demo"
+    )
+
+    package.add_dependency("demo", {"path": path, "extras": ["foo"]})
+
+    ops = solver.solve()
+
+    check_solver_result(
+        ops,
+        [
+            {"job": "install", "package": cleo},
+            {"job": "install", "package": pendulum},
+            {"job": "install", "package": get_package("demo", "0.1.2")},
+        ],
+    )
+
+    op = ops[2]
+
+    assert op.package.name == "demo"
+    assert op.package.version.text == "0.1.2"
+    assert op.package.source_type == "directory"
+    assert op.package.source_url == path
+
+
+def test_solver_can_resolve_sdist_dependencies(solver, repo, package):
+    pendulum = get_package("pendulum", "2.0.3")
+    repo.add_package(pendulum)
+
+    path = str(
+        Path(__file__).parent.parent
+        / "fixtures"
+        / "distributions"
+        / "demo-0.1.0.tar.gz"
+    )
+
+    package.add_dependency("demo", {"path": path})
+
+    ops = solver.solve()
+
+    check_solver_result(
+        ops,
+        [
+            {"job": "install", "package": pendulum},
+            {"job": "install", "package": get_package("demo", "0.1.0")},
+        ],
+    )
+
+    op = ops[1]
+
+    assert op.package.name == "demo"
+    assert op.package.version.text == "0.1.0"
+    assert op.package.source_type == "file"
+    assert op.package.source_url == path
+
+
+def test_solver_can_resolve_sdist_dependencies_with_extras(solver, repo, package):
+    pendulum = get_package("pendulum", "2.0.3")
+    cleo = get_package("cleo", "1.0.0")
+    repo.add_package(pendulum)
+    repo.add_package(cleo)
+
+    path = str(
+        Path(__file__).parent.parent
+        / "fixtures"
+        / "distributions"
+        / "demo-0.1.0.tar.gz"
+    )
+
+    package.add_dependency("demo", {"path": path, "extras": ["foo"]})
+
+    ops = solver.solve()
+
+    check_solver_result(
+        ops,
+        [
+            {"job": "install", "package": cleo},
+            {"job": "install", "package": pendulum},
+            {"job": "install", "package": get_package("demo", "0.1.0")},
+        ],
+    )
+
+    op = ops[2]
+
+    assert op.package.name == "demo"
+    assert op.package.version.text == "0.1.0"
+    assert op.package.source_type == "file"
+    assert op.package.source_url == path
+
+
+def test_solver_can_resolve_wheel_dependencies(solver, repo, package):
+    pendulum = get_package("pendulum", "2.0.3")
+    repo.add_package(pendulum)
+
+    path = str(
+        Path(__file__).parent.parent
+        / "fixtures"
+        / "distributions"
+        / "demo-0.1.0-py2.py3-none-any.whl"
+    )
+
+    package.add_dependency("demo", {"path": path})
+
+    ops = solver.solve()
+
+    check_solver_result(
+        ops,
+        [
+            {"job": "install", "package": pendulum},
+            {"job": "install", "package": get_package("demo", "0.1.0")},
+        ],
+    )
+
+    op = ops[1]
+
+    assert op.package.name == "demo"
+    assert op.package.version.text == "0.1.0"
+    assert op.package.source_type == "file"
+    assert op.package.source_url == path
+
+
+def test_solver_can_resolve_wheel_dependencies_with_extras(solver, repo, package):
+    pendulum = get_package("pendulum", "2.0.3")
+    cleo = get_package("cleo", "1.0.0")
+    repo.add_package(pendulum)
+    repo.add_package(cleo)
+
+    path = str(
+        Path(__file__).parent.parent
+        / "fixtures"
+        / "distributions"
+        / "demo-0.1.0-py2.py3-none-any.whl"
+    )
+
+    package.add_dependency("demo", {"path": path, "extras": ["foo"]})
+
+    ops = solver.solve()
+
+    check_solver_result(
+        ops,
+        [
+            {"job": "install", "package": cleo},
+            {"job": "install", "package": pendulum},
+            {"job": "install", "package": get_package("demo", "0.1.0")},
+        ],
+    )
+
+    op = ops[2]
+
+    assert op.package.name == "demo"
+    assert op.package.version.text == "0.1.0"
+    assert op.package.source_type == "file"
+    assert op.package.source_url == path