+++ poetry/masonry/builders/wheel.py
@@ -11,6 +11,7 @@ import zipfile
 
 from base64 import urlsafe_b64encode
 from io import StringIO
+from typing import Set
 
 from poetry.__version__ import __version__
 from poetry.semver import parse_constraint
@@ -185,9 +186,9 @@ class WheelBuilder(Builder):
             # RECORD itself is recorded with no hash or size
             f.write(self.dist_info + "/RECORD,,\n")
 
-    def find_excluded_files(self):  # type: () -> list
+    def find_excluded_files(self):  # type: () -> Set
         # Checking VCS
-        return []
+        return set()
 
     @property
     def dist_info(self):  # type: () -> str