+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed reading of some `setup.py` files.
 - Fixed a `KeyError` when getting information for packages which require reading setup files.
 - Fixed the building of wheels with C extensions and an `src` layout.
+- Fixed extras being selected when resolving dependencies even when not required.
 
 
 ## [0.12.6] - 2018-11-05