+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed a `KeyError` when getting information for packages which require reading setup files.
 - Fixed the building of wheels with C extensions and an `src` layout.
 - Fixed extras being selected when resolving dependencies even when not required.
+- Fixed performance issues when packaging projects if a lot of files were excluded.
 
 
 ## [0.12.6] - 2018-11-05